+++ embassy-usb/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `UAC1`: Add audio source
 - `UAC1`: `Speaker::new` now returns `Self` with the parts inside instead of a tuple
 - `CDC-NCM`: Handle `SetEthernetPacketFilter` and advertise it in `bmNetworkCapabilities`, which also works around a macOS bug that intermittently left the data interface disabled
+- `MIDI`: Allow sender-only or receiver-only configuration
 
 ## 0.6.0 - 2026-03-10